requirements.txt
anyio==4.10.0
aplr==10.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_cytoscape==1.0.2
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
folium==0.20.0
fonttools==4.59.0
fqdn==1.5.1
gevent==25.8.2
graphviz==0.21
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
interpret-core==0.7.2
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.3
mlxtend==0.23.4
multiprocess==0.70.18
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
rsa==4.9.1
SALib==1.5.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.5
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.8.28
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
xyzservices==2025.4.0
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
-e git+https://github.com/Tenshin000/CogniPredictAD@09d06d857cf66b6ea2cdd545efa358ba0abddd68#egg=CogniPredictAD
